refactor(mcp): move connection status signal to base class
Remove vscode-specific McpConnection wrapper and move status signal logic to the common base class. This simplifies the architecture by eliminating redundant code and centralizing status management.
